Remove @trusted function attributes
The @trusted attribute is used for memory safety.  It allows the programmer to provide a @safe interface, but use unsafe code in the body.  Essentially the programmer is telling the compiler "trust" me, this function is safe even though I'm using system features.

It is not relevant to the spirit of this benchmark, and appears to have no affect on the performance.